Make find of data-coolclock-* attributes more stable.
The DOM dataset object is every data-* attribute with hyphenated
gaps converted to camelCase. This code was making assumptions of
how the user would write the data attributes and how it would end
up as camelCase.

Now we lowercase the dataset property and see if it is a property
of the optDataset object. If it is, the corresponding value is the
correct camelCased opt name.