Avoid setting options that weren't passed when finding clocks.
Previously, after getting all the options that were set, the
settings for some options were being set regardless of them being
set in the html. This could override the default in CoolClock.config
when init() is run. Even if it ends up being the same as the default,
it would still be duplicating the determination of the default,
which is totes bad guise, srsly.