Determine boolean settings correctly. Allow noSeconds override.
13516671d649944e29f67741e2a694bbdf01e4f0 resulted in some settings
not being set because it depended on the order of the loop. Silly
Henry.

Now duplicated code has been reduced, settings are being checked
when they are all there, and noSeconds can correctly override
showSecondHand and showDigitalSecs.

noSeconds performs the override before the clocks are initiated by
changing the value of showSecondHand and showDigitalSecs, meaning
we don't have to add more checks into the prototype code.